pyproject.toml
[project]
name = "aion1-experiments"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "astropy>=7.1.1",
    "datasets>=4.4.1",
    "huggingface-hub>=1.1.5",
    "ipykernel>=7.1.0",
    "lightning>=2.5.6",
    "matplotlib>=3.10.7",
    "pandas>=2.3.3",
    "polymathic-aion>=0.0.2",
    "pyarrow>=22.0.0",
    "python-dotenv>=1.2.1",
    "safetensors>=0.7.0",
    "scienceplots>=2.2.0",
    "seaborn>=0.13.2",
    "torch",
]
